TST: modify flaky test for constrained minimization.
Was failing with:
```
lib/python3.9/site-packages/scipy/optimize/_hessian_update_strategy.py:182: in update
    warn('delta_grad == 0.0. Check if the approximated '
E   UserWarning: delta_grad == 0.0. Check if the approximated function is linear. If the function is linear better results can be obtained by defining the Hessian as zero instead of using quasi-Newton approximations.
```

Follow-up to gh-12922, where the `grad_delta==0` warning was observed,
but apparently the fix was a little unstable (seeing it in CI of the
Meson build on my fork).